Update Node.js version and TunnelHub SDK
This commit updates the Node.js runtime version from 18.x to 20.x in the `tunnelhub.yml` and `.nvmrc` files. It also upgrades the TunnelHub SDK version from 1.3.4 to 1.3.5 in `package.json` and updates the references accordingly in `yarn.lock` and `integration.ts`. Additionally, the `data.d.ts` file has been renamed to `types.ts` to reflect changes made in the project structure.